实装：（绘制）增加了指令数值结算信息中的打印信息上色，包括角色的口上颜色，体力、气力、理智、精液等 调整：优化了富文本代码文件和调用富文本的各代码文件中对于富文本的调用方式 调整：将角色头部信息面板从原文件中独立出来作为单独的文件进行处理了

+++ Script/Core/text_handle.py
@@ -102,7 +102,7 @@ def id_index(now_id: int) -> str:
     return f"[{str(now_id).zfill(3)}]"
 
 
-def number_to_symbol_string(value: int) -> str:
+def number_to_symbol_string(value: float) -> str:
     """
     数字转换为带正负数符号的数字字符串
     Keyword arguments:
